List Component Code Enhancement
